Rename modules to System.Metrics.Prometheus
