+++ test/State.hs
@@ -20,7 +20,7 @@ import qualified Test.QuickCheck as QC
 import qualified Test.SmallCheck as SC
 import qualified Test.SmallCheck.Series as SC
 
-import System.Metrics.Internal.State
+import System.Metrics.Prometheus.Internal.State
 
 ------------------------------------------------------------------------
